committed

+++ src/main/java/com/jovan/erp_v1/service/FileOptService.java
@@ -1,6 +1,8 @@
 package com.jovan.erp_v1.service;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
@@ -9,6 +11,8 @@ import org.springframework.transaction.annotation.Transactional;
 import com.jovan.erp_v1.enumeration.FileAction;
 import com.jovan.erp_v1.enumeration.FileExtension;
 import com.jovan.erp_v1.exception.FileOptErrorException;
+import com.jovan.erp_v1.exception.NoDataFoundException;
+import com.jovan.erp_v1.exception.ValidationException;
 import com.jovan.erp_v1.mapper.FileOptMapper;
 import com.jovan.erp_v1.model.FileOpt;
 import com.jovan.erp_v1.repository.FileOptRepository;
@@ -27,6 +31,7 @@ public class FileOptService implements IFileOptService {
     @Transactional
     @Override
     public FileOptResponse create(FileOptRequest request) {
+    	validateFileOptRequest(request);
         FileOpt fileOpt = fileOptMapper.toEntity(request);
         fileOptRepository.save(fileOpt);
         return fileOptMapper.toResponse(fileOpt);
@@ -40,6 +45,7 @@ public class FileOptService implements IFileOptService {
 		}
         FileOpt fileOpt = fileOptRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("FileOpt not found"));
+        validateFileOptRequest(request);
         fileOpt.setExtension(request.extension());
         fileOpt.setMimeType(request.mimeType());
         fileOpt.setMaxSizeInBytes(request.maxSizeInBytes());
@@ -53,6 +59,9 @@ public class FileOptService implements IFileOptService {
     @Transactional
     @Override
     public void delete(Long id) {
+    	if(!fileOptRepository.existsById(id)) {
+    		throw new NoDataFoundException("FileOPT is not found "+id);
+    	}
         fileOptRepository.deleteById(id);
     }
 
@@ -67,7 +76,11 @@ public class FileOptService implements IFileOptService {
     @Override
     @Transactional(readOnly = true)
     public List<FileOptResponse> getAll() {
-        return fileOptRepository.findAll()
+    	List<FileOpt> items = fileOptRepository.findAll();
+    	if(items.isEmpty()) {
+    		throw new NoDataFoundException("List of FileOPT is empty");
+    	}
+        return items
                 .stream()
                 .map(fileOptMapper::toResponse)
                 .toList();
@@ -76,7 +89,13 @@ public class FileOptService implements IFileOptService {
     @Override
     @Transactional(readOnly = true)
     public List<FileOptResponse> getByExtension(FileExtension extension) {
-        return fileOptRepository.findByExtension(extension)
+    	validateFileExtension(extension);
+    	List<FileOpt> items = fileOptRepository.findByExtension(extension);
+    	if(items.isEmpty()) {
+    		String msg = String.format("No FileOPT for file extension %s is found", extension);
+    		throw new NoDataFoundException(msg);
+    	}
+        return items
                 .stream()
                 .map(fileOptMapper::toResponse)
                 .toList();
@@ -85,8 +104,73 @@ public class FileOptService implements IFileOptService {
     @Override
     @Transactional(readOnly = true)
     public List<FileOptResponse> getByAction(FileAction action) {
-        return fileOptRepository.findByAvailableActions(action)
+    	validateFileAction(action);
+    	List<FileOpt> items = fileOptRepository.findByAvailableActions(action);
+    	if(items.isEmpty()) {
+    		String msg = String.format("No FileOpt for file action %s is found", action);
+    		throw new NoDataFoundException(msg);
+    	}
+        return items
                 .stream().map(fileOptMapper::toResponse)
                 .collect(Collectors.toList());
+    }
+    
+    private void validateFileExtension(FileExtension extension) {
+    	Optional.ofNullable(extension)
+    		.orElseThrow(() -> new ValidationException("FileExtension extension must not be null"));
+    }
+    
+    private void validateFileAction(FileAction action) {
+    	Optional.ofNullable(action)
+    		.orElseThrow(() -> new ValidationException("FileAction action must not be null"));
+    }
+    
+    private void validateFileOptRequest(FileOptRequest request) {
+    	if(request == null) {
+    		throw new ValidationException("FileOptRequest request must not be null");
+    	}
+    	validateFileExtension(request.extension());
+    	validateString(request.mimeType());
+    	validateLong(request.maxSizeInBytes());
+    	validateBoolean(request.uploadEnabled());
+    	validateBoolean(request.previewEnabled());
+    	validateFileAction(request.availableActions());
+    }
+    
+    private void validateString(String str) {
+    	if(str == null || str.trim().isEmpty()) {
+    		throw new ValidationException("String must not be null nor empty");
+    	}
+    }
+    
+    private void validateFileAction(Set<FileAction> availableActions) {
+    	if(availableActions == null || availableActions.isEmpty()) {
+    		throw new ValidationException("FileAction must not be empty nor null");
+    	}
+    	for(FileAction fa : availableActions) {
+    		if(fa == null) {
+    			throw new ValidationException("FileAction must not be null");
+    		}
+    		validateFileActionRequest(fa);
+    	}
+    }
+    
+    private void validateFileActionRequest(FileAction act) {
+    	if(act == null) {
+    		throw new ValidationException("FileAction must not be null");
+    	}
+    	validateFileAction(act);
+    }
+    
+    private void validateLong(Long num) {
+    	if(num == null || num <= 0) {
+    		throw new ValidationException("Number must not be null nor negative");
+    	}
+    }
+    
+    private void validateBoolean(Boolean bool) {
+    	if(bool == null) {
+    		throw new ValidationException("Boolean value must not be null");
+    	}
     }
 }
